move get_data_shape and fix imports
